animation system test

+++ source/DxlibMario/RigidBody2D.cpp
@@ -52,7 +52,7 @@ void RigidBody2D::ComputeFinalPosition(int & x, int & y, const Rect & rect1, con
 	if (_velocity.x != 0)
 	{
 		Rect tmpRect(rect1);
-		tmpRect.Shift(_gameobject->GetGlobalPosition().x + _velocity.x, _gameobject->GetGlobalPosition().y);
+		tmpRect.Shift((int)_gameobject->GetGlobalPosition().x + (int)_velocity.x, (int)_gameobject->GetGlobalPosition().y);
 		if (Rect::IsCollision(tmpRect, rect2)) {
 			if (_velocity.x > 0) {
 				int tmp = rect2._left - 1 - rect1._right;
@@ -72,7 +72,7 @@ void RigidBody2D::ComputeFinalPosition(int & x, int & y, const Rect & rect1, con
 	if (_velocity.y != 0)
 	{
 		Rect tmpRect(rect1);
-		tmpRect.Shift(_gameobject->GetGlobalPosition().x, _gameobject->GetGlobalPosition().y + _velocity.y);
+		tmpRect.Shift((int)_gameobject->GetGlobalPosition().x, (int)_gameobject->GetGlobalPosition().y + (int)_velocity.y);
 		if (Rect::IsCollision(tmpRect, rect2)) {
 			if (_velocity.y > 0) {
 				int tmp = rect2._up - 1 - rect1._down;
